python_version/requirements.txt
cryptography==41.0.5
firebase-admin==6.2.0
Flask-Cors==4.0.0
grpcio-status==1.59.2
python-dotenv==1.0.0
setuptools==58.